requirements.txt
streamlit 
pillow 
pytesseract 
gtts
